Extract HTTP schemas/constants/helpers into flux/api/schemas.py
First step of the server.py decomposition. Moves the ~30 request/response
DTOs, the MAX_WORKFLOW_UPLOAD_BYTES / SERVICE_NAME_RE constants, and the
_rate_limit_exceeded_handler / _has_any_workflow_read / _inject_trace_context
helpers out of server.py into a dependency-free module the per-domain route
modules can share. Re-exported from flux.server for backward compatibility.
No behavior change; full unit suite green (2263 passed).
